Fix featuretests, yay
